+++ front/components/Modal/ModalContent.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const ModalContent = () => {
+  return (
+    <div>
+      <div>모달 컨텐츠</div>
+      <div>모달 컨텐츠</div>
+    </div>
+  );
+};
+
+export default ModalContent;
