+++ front/styles/Modal/modalCarousel.ts
